+++ app/controllers/skaters_controller.rb
@@ -18,13 +18,9 @@ class SkatersController < ApplicationController
   ################################################################
   ## show
   def show
-    show_skater(Skater.find_by(isu_number: params[:isu_number]))
-  end
-  def show_by_name
-    show_skater(Skater.find_by(name: params[:name]))
-  end
-  def show_skater(skater)
-    raise ActiveRecord::RecordNotFound.new("no such skater") if skater.nil?
+    skater = Skater.find_by(isu_number: params[:isu_number]) ||
+      Skater.find_by(name: params[:isu_number]) || 
+      raise(ActiveRecord::RecordNotFound.new("no such skater"))
 
     ################
     ## competition results