+++ app/decorators/competition_decorator.rb
@@ -8,8 +8,8 @@ end
 
 class CompetitionDecorator < EntryDecorator
   def name
-    n = h.link_to_competition(model)
-    (model.isu_championships) ? h.content_tag(:b, n) : n
+    h.link_to_competition(model)
+    #(model.isu_championships) ? h.content_tag(:b, n) : n
   end  
   def site_url
     h.link_to_competition_site("Official", model)