+++ lib/fisk8_result_viewer/updater/competition_updater.rb
@@ -1,3 +1,5 @@
+using Seniorize
+
 module Fisk8ResultViewer
   module Updater
     class CompetitionUpdater